requirements.txt
asgiref==3.8.1
Django==5.1.2
django-mssql-backend==2.8.1
djangorestframework==3.15.2
pyodbc==5.1.0
sqlparse==0.5.1
